Fixed a bug in Communicator where Isend was being used instead of send. This caused issues when events happened to be sent in the following order:
   -  Isend(event)
   -  delete[] (char*) event

Isend requires event to be preserved until it is completely sent out. 
In the above situation the buffer was being deleted even before MPI finished
sending out the event. This causes all kinds of wierd problems. I am surprised
simulations were even running this far.

there is another weird condition that I noticed --- there is an anti-message
for which a suitable message in the input queue does not exist. This seems
like a weird situation. Currently, I added code to abort if such a situation
arises during simulation-time.

+++ kernel/include/f_heap.h
@@ -80,9 +80,9 @@ protected:
 
   void ppHelper(std::ostream& os, const std::string &indent) const{
       os << indent <<((is_root())? "*" : ">") << *data() << std::endl;
-      for (int i=0; i <  m_children.size(); i++) {
+      for (size_t i=0; (i <  m_children.size()); i++) {
           if ( m_children[i] != 0 ){
-              m_children[i]->ppHelper(os,indent+"-");
+              m_children[i]->ppHelper(os, indent + "-");
           }
       }
   }
@@ -139,7 +139,7 @@ public:
   iterator end() const   { return iterator(this, 0); }
 
   void prettyPrint(std::ostream& os) const {
-      for (int i=0; i < m_roots.size(); i++){
+      for (size_t i=0; (i < m_roots.size()); i++){
           if (m_roots[i] != 0 ){
               m_roots[i]->ppHelper(os,"-");
           }
